Rectangle Joint abstract class for knee and elbow
