+++ src/joint.cpp
@@ -17,7 +17,6 @@ void f3(float * x, float * fvec, float ** fjac, Plane * plane1, Plane * plane2, 
     float a = x[0];
     float b = x[1];
     float t = x[2];
-    //std::cout << *plane1 << *plane2 << std::endl;
     auto n_1 = plane1->getN(); auto n_2 = plane2->getN();
     float nx_1 = n_1.getDirection().x();
     float ny_1 = n_1.getDirection().y();
@@ -121,108 +120,25 @@ void f3(float * x, float * fvec, float ** fjac, Plane * plane1, Plane * plane2, 
 }
 
 /*!
- * \brief The BezierCoords3D class - координаты Безье (альфа, бета, t)
- */
-class BezierCoords3D
-{
-public:
-    BezierCoords3D(): m_alpha{-0.1}, m_beta{-0.1}, m_t{-1.0}
-    {}
-
-    BezierCoords3D(double a,double b, double t): m_alpha{a}, m_beta{b}, m_t{t}
-    {}
-
-    ~BezierCoords3D()
-    {}
-
-    double const & alpha() const { return m_alpha; }
-    double const & beta() const { return m_beta; }
-    double const & t() const { return m_t; }
-    double m_alpha;
-    double m_beta;
-    double m_t;
-};
-
-/*!
- * \brief StartCoefs - ищет начальное приближение для (альфа, бета, t)
- * Для оценки альфа использует оценку снизу - опускает перпендикуляры из точки на плоскости
- * на каждой плоскости находит альфы для проекций точки
- * выбирает наибольшее из двух.
- * Аналогично ищет бета.
- * t  ищет из отношения перпендикуляров, с учётом их знака:
- * t = (дл.перп. к пл. 1)/(длина перпендикуляра к плоск.1 + длина п. к пл.2)
- * \param plane1
- * \param plane2
- * \param point
- * \return - начальные (альфа, бета, t)
- */
-BezierCoords3D * StartCoefs(Plane * plane1, Plane * plane2, Point3D<float> * point)
-{
-    auto n1 = plane1->getN();
-    auto dir1 = n1.getDirection();
-    float A1 = dir1.x(); float B1 = dir1.y(); float C1 = dir1.z();
-    auto e1_1 = plane1->getE1(); auto e2_1 = plane1->getE2();
-    auto ver1 = e1_1.getPosition();
-    float D1 = -ver1.x()*A1 - ver1.y()*B1 - ver1.z()*C1;
-    float lenn1_s = (A1*point->x() + B1*point->y() + C1*point->z() + D1) / sqrt(A1*A1+B1*B1+C1*C1);
-    if (lenn1_s < 0) {
-        return new BezierCoords3D(1,1,-1);
-    }
-    float lenn1 = fabs(lenn1_s);
-    Point3D <float> pn1 = {point->x() - A1*lenn1, point->y() - B1*lenn1, point->z() - C1*lenn1};
-    Vector3D v1 = {true, ver1, pn1};
-    float a1 = (v1.getDirection().x()*e1_1.getDirection().x() +
-            v1.getDirection().y()*e1_1.getDirection().y() +
-            v1.getDirection().z()*e1_1.getDirection().z()) * v1.getLength()/e1_1.getLength();
-    float b1 = (v1.getDirection().x()*e2_1.getDirection().x() +
-            v1.getDirection().y()*e2_1.getDirection().y() +
-            v1.getDirection().z()*e2_1.getDirection().z()) * v1.getLength()/e2_1.getLength();
-
-    auto n2 = plane2->getN();
-    auto dir2 = n2.getDirection();
-    float A2 = dir2.x(); float B2 = dir2.y(); float C2 = dir2.z();
-    auto e1_2 = plane2->getE2(); auto e2_2 = plane2->getE1();
-    auto ver2 = e1_2.getPosition();
-    float D2 = -ver2.x()*A2 - ver2.y()*B2 - ver2.z()*C2;
-    float lenn2_s = (A2*point->x() + B2*point->y() + C2*point->z() + D2) / sqrt(A2*A2+B2*B2+C2*C2);
-    if (lenn2_s < 0) {
-        return new BezierCoords3D(1,1,-1);
-    }
-    float lenn2 = fabs(lenn2_s);
-    Point3D <float> pn2 = {point->x() - A2*lenn2, point->y() - B2*lenn2, point->z() - C2*lenn2};
-    Vector3D v2 = {true, ver2, pn2};
-    float a2 = (v2.getDirection().x()*e1_2.getDirection().x() +
-            v2.getDirection().y()*e1_2.getDirection().y() +
-            v2.getDirection().z()*e1_2.getDirection().z()) * v2.getLength()/e1_2.getLength();
-    float b2 = (v2.getDirection().x()*e2_2.getDirection().x() +
-            v2.getDirection().y()*e2_2.getDirection().y() +
-            v2.getDirection().z()*e2_2.getDirection().z()) * v2.getLength()/e2_2.getLength();
-    //std::cout <<"\n"<< a1 << " " << a2 << " " << b1 << " " << b2 << " " << lenn1_s << " " << lenn2_s << std::endl;
-    //std::cout <<*plane1 << *plane2 <<v1<<v2<< std::endl;
-//    return new BezierCoords3D(a1 > a2 ? a2 : a1, b1 > b2 ? b1 : b2, lenn1_s/(lenn1_s+lenn2_s));
-    return new BezierCoords3D(lenn1 > lenn2 ? a2 : a1, lenn1 > lenn2 ? b2 : b1, lenn1_s/(lenn1_s+lenn2_s));
-}
-
-/*!
- * \brief FindAlpha - для точки с известными дек. координатами ищет координаты Безье
- * \param plane1 - первая плоскость, альфа вдоль E1, бета вдоль E2
- * \param plane2 - вторая плоскость, альфа вдоль E2, бета вдоль E1
+ * \brief findAlpha - для точки с известными дек. координатами ищет координаты Безье
+ * в деформированном суставе
  * \param point - точка (координаты Безье)
  * \param der - допустимое отклонение дек.координат (сумма квадратов)
  * \return - (альфа, бета, t)
  */
-BezierCoords3D * FindAlpha(Plane * plane1, Plane * plane2, Point3D<float> * point, float der)
+BezierCoords3D * Joint::findAlpha(Point3D<float> * point, float der)
 {
-    auto startb = StartCoefs(plane1,plane2,point);
+    auto startb = this->startCoefs(point);
     float a = startb->alpha();
     float b = startb->beta();
     float t = startb->t();
     if (t < 0)
         return startb;
-    std::function<void(float * , int , float * , float ** )> func = [plane1, plane2, point]
+    auto plane1 = &m_endPlane1; auto plane2 = &m_endPlane2;
+    std::function<void(float * , int , float * , float ** )> func = [plane1 ,plane2, point]
             (float * x, int n, float * fvec, float ** fjac)
     {
-        f3(x,fvec,fjac,plane1, plane2,point);
+        f3(x,fvec,fjac,plane1 ,plane2,point);
     };
     float x[3];
     x[0] = a;
@@ -235,25 +151,23 @@ BezierCoords3D * FindAlpha(Plane * plane1, Plane * plane2, Point3D<float> * poin
 }
 
 /*!
- * \brief FindPoint3 - считает координаты точки из координат Безте
- * \param plane1 - первая плоскость, альфа вдоль E1, бета вдоль E2
- * \param plane2 - вторая плоскость, альфа вдоль E2, бета вдоль E1
+ * \brief findPoint - считает координаты точки из координат Безье в недеформированном суставе
  * \param bc - (альфа, бета, t)
  * \return - точка в декартовых координатах
  */
-Point3D<float> * FindPoint(Plane * plane1, Plane * plane2, BezierCoords3D * bc)
+Point3D<float> * Joint::findPoint(BezierCoords3D * bc)
 {
     float a = bc->alpha();
     float b = bc->beta();
     float t = bc->t();
-    auto n_1 = plane1->getN(); auto n_2 = plane2->getN();
+    auto n_1 = m_startPlane1.getN(); auto n_2 = m_startPlane2.getN();
     float nx_1 = n_1.getDirection().x();
     float ny_1 = n_1.getDirection().y();
     float nz_1 = n_1.getDirection().z();
     float nx_2 = n_2.getDirection().x();
     float ny_2 = n_2.getDirection().y();
     float nz_2 = n_2.getDirection().z();
-    auto e1_1 = plane1->getE1(); auto e2_1 = plane1->getE2();
+    auto e1_1 = m_startPlane1.getE1(); auto e2_1 = m_startPlane1.getE2();
     float e1x_1 = e1_1.getDirection().x() * e1_1.getLength();
     float e1y_1 = e1_1.getDirection().y() * e1_1.getLength();
     float e1z_1 = e1_1.getDirection().z() * e1_1.getLength();
@@ -262,7 +176,7 @@ Point3D<float> * FindPoint(Plane * plane1, Plane * plane2, BezierCoords3D * bc)
     float e2z_1 = e2_1.getDirection().z() * e2_1.getLength();
     auto p0_1 = e1_1.getPosition();
     float x0_1 = p0_1.x(); float y0_1 = p0_1.y(); float z0_1 = p0_1.z();
-    auto e2_2 = plane2->getE1(); auto e1_2 = plane2->getE2();
+    auto e2_2 = m_startPlane2.getE1(); auto e1_2 = m_startPlane2.getE2();
     float e1x_2 = e1_2.getDirection().x() * e1_2.getLength();
     float e1y_2 = e1_2.getDirection().y() * e1_2.getLength();
     float e1z_2 = e1_2.getDirection().z() * e1_2.getLength();
@@ -299,10 +213,8 @@ Point3D<float> * FindPoint(Plane * plane1, Plane * plane2, BezierCoords3D * bc)
 
 bool Joint::getStartPoint(Point3D <float> * end, Point3D <float> * start, float der)
 {
-    //std::cout << m_startPlane2 << m_endPlane2;
-    //exit(0);
-    auto bezier = FindAlpha(&m_endPlane1, &m_endPlane2, end, der);
-    auto point = FindPoint(&m_startPlane1, &m_startPlane2, bezier);
+    auto bezier = this->findAlpha(end, der);
+    auto point = this->findPoint(bezier);
     *start = *point;
     if (bezier->t() < 0 || bezier->t() > 1)
     {
